+++ src/test/kotlin/in/kkkev/jjidea/vcs/changes/JujutsuChangeProviderTest.kt
@@ -580,6 +580,20 @@ class JujutsuChangeProviderTest {
     }
 
     @Test
+    fun `collectTrackedAbsolutePaths - single file does not include summary lines`() {
+        val output = statusOutput("M foo.txt")
+        val paths = jcp.collectTrackedAbsolutePaths(output, repo)
+        paths shouldBe setOf("${directory.path}/foo.txt")
+    }
+
+    @Test
+    fun `collectTrackedAbsolutePaths - multiple files do not include Parent commit line`() {
+        val output = statusOutput("M foo.txt", "A bar/baz.kt", "D old.txt")
+        val paths = jcp.collectTrackedAbsolutePaths(output, repo)
+        paths shouldBe setOf("${directory.path}/foo.txt", "${directory.path}/bar/baz.kt", "${directory.path}/old.txt")
+    }
+
+    @Test
     fun `explicit conflict paths from resolveList override warning section`() {
         val longPath = "gateway/lib/gatewaysnapshotvolatileservice.go"
         directory.addChild(getOrCreateVirtualFile(false, longPath))